test: move request logging wiring tests to integration
